SS8-150
Set up YOLOv7 and Intel OpenVino environments locally on a Mac computer for testing of the different models on speed and accuracy.
Acceptance Criteria: Create a file with the speed and accuracy numbers of the YOLOv7 and IR models. Create a bar chart displaying the differences between the models as well as the machines.